requirements.txt
flask>=2.2.5
Flask-RESTful==0.3.9
SQLAlchemy==1.4.25
mysql-connector-python==8.0.28
Flask-Cors
